Change view_model.h to use size_t.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710687
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017319}

+++ ash/app_list/views/apps_grid_view_test_api.h
@@ -49,7 +49,7 @@ class AppsGridViewTestApi {
 
   void PressItemAt(int index);
 
-  int TilesPerPage(int page) const;
+  size_t TilesPerPage(int page) const;
 
   int AppsOnPage(int page) const;
